Fix circular dependency issues by consolidating files
